pytool/requirements.txt
bottle==0.12.21
requests
psutil
mysql-connector-python
xlrd
pandas
ffmpy
qrcode[pil]
kafka-python
python-snappy
pycron
tornado
pyjwt
redis
ecdsa
base58
pyminizip
webssh
dnspython